% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bat_file.R
\name{make_bat_file}
\alias{make_bat_file}
\title{.bat shortcut for monitorgonio}
\usage{
make_bat_file(outfile = "~/Desktop/monitorgonio.bat")
}
\arguments{
\item{ouputdir}{where to save the .bat file}
}
\description{
automatically generates a .bat which should load monitorgonio successfully on windows
}
