% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sh_file.R
\name{make_sh_file}
\alias{make_sh_file}
\title{.sh shortcut for monitorgonio}
\usage{
make_sh_file(outfile = "~/Desktop/monitorgonio.sh")
}
\arguments{
\item{outfile}{where to save the .sh file}
}
\description{
automatically generates a .sh which should load monitorgonio successfully on mac / linux
}
