% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitorgonio.R
\docType{package}
\name{monitorgonio-package}
\alias{monitorgonio}
\alias{monitorgonio-package}
\title{monitorgonio: a quick hack for visualizing Argos Goniometer output in the field}
\description{
WARNING: this is still very rough. use at your own risk. Just a cobbled together shiny app for visualizing the direction and strength of a ptt using a laptop connected to an Argos Goniometer. It just easedrops on the log file. useful because the little handheld can be a pain to read on a bouncing boat.
}
\details{
you probably want to start by looking at \code{\link[monitorgonio]{run_monitorgonio}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://williamcioffi.github.io/monitorgonio}
  \item Report bugs at \url{https://github.com/williamcioffi/monitorgonio/issues}
}

}
\author{
\strong{Maintainer}: William R. Cioffi \email{wrc14@duke.edu}

}
\keyword{internal}
