% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pttkey.R
\docType{data}
\name{pttkey}
\alias{pttkey}
\title{an example pttkey}
\format{a data.frame}
\usage{
data(pttkey)
}
\description{
a tiny example pttkey which also works with the simulated goniometer output see\code{\link[monitorgonio]{simulate_gonio}}.
}
\keyword{datasets}
