% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_gonio.R
\name{simulate_gonio}
\alias{simulate_gonio}
\title{simulate some gonio hits}
\usage{
simulate_gonio(logfile)
}
\description{
simulate some gonio hits by spitting lines to a file a couple of seconds apart
}
\examples{
\dontrun{
# in one instance of r start the shiny app
# run_monitorgonio()

# start a new isntance of R
# write the example pttkey
write.csv(pttkey, file = "pttkey.csv", row.names = FALSE, sep = ',')
# make an empty log file
cat("", file = "log")
# load the log file and pttkey into the shiny app
# fill the log file
simualate_gonio("log")
}
}
