% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitorgonio.R
\docType{package}
\name{monitorgonio-package}
\alias{monitorgonio}
\alias{monitorgonio-package}
\title{monitorgonio: visualize Argos Goniometer output in the field}
\description{
A shiny app for visualizing the direction and strength of a ptt using a laptop connected to an Argos Goniometer. It just eavesdrops on the log file. Can be useful because the little handheld can be a pain to read on a bouncing boat.
}
\details{
you probably want to start by looking at \code{\link[monitorgonio]{run_monitorgonio}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/williamcioffi/monitorgonio}
  \item Report bugs at \url{https://github.com/williamcioffi/monitorgonio/issues}
}

}
\author{
\strong{Maintainer}: William R. Cioffi \email{wrc14@duke.edu} (\href{https://orcid.org/0000-0003-1182-8578}{ORCID})

}
\keyword{internal}
